services/qdb/requirements.txt
aiohttp
gunicorn
numpy
protobuf
pycryptodome